Use venusian to import task
venusian is usesull to be less intrusive
in task declaration. Until the configurator
is used, the decorator do nothing. This is
great for testing tasks code.

note that the task.install method dynamically
patch the module to install the decorator,
if task has been imported becore configuration
made, they are not patched...